Recombine some files & some cleanup
